Move Skia related files to the new folder
